Changed inheritance + removed security level
By josef and anders
