+++ src/Domain/CaseEmployee.java
@@ -1,0 +1,22 @@
+/*
+ * To change this license header, choose License Headers in Project Properties.
+ * To change this template file, choose Tools | Templates
+ * and open the template in the editor.
+ */
+package Domain;
+
+/**
+ *
+ * @author ander
+ */
+public class CaseEmployee extends Employee{
+
+    public CaseEmployee(long cpr, String name, char gender, String birthDate, String address, int phoneNumber, String mail, int id, String userName, String password) {
+        super(cpr, name, gender, birthDate, address, phoneNumber, mail, id, userName, password);
+    }
+    
+    public void  createCaseRequest() {
+        
+    }
+    
+}
