+++ src/Domain/Employee.java
@@ -11,18 +11,16 @@ import java.util.Date;
  *
  * @author ander
  */
-public class Employee extends Citizen implements IEmployee{
+public class Employee extends Person implements IEmployee{
     private int id; 
     private String userName;
     private String password; 
-    private int securityLevel; 
     
-    public Employee(long cpr, String name, char gender, String birthDate, String address, int phoneNumber, String mail, int id, String userName, String password, int securityLevel) {
+    public Employee(long cpr, String name, char gender, String birthDate, String address, int phoneNumber, String mail, int id, String userName, String password) {
         super(cpr, name, gender, birthDate, address, phoneNumber, mail);
         this.id = id;
         this.userName = userName;
         this.password = password;
-        this.securityLevel = securityLevel;
     }
     
     @Override
@@ -41,11 +39,6 @@ public class Employee extends Citizen implements IEmployee{
     }
     
     @Override
-    public int getSecurityLevel() {
-        return securityLevel;
-    }
-
-    @Override
     public void setId(int id) {
         this.id = id;
     }
@@ -58,10 +51,5 @@ public class Employee extends Citizen implements IEmployee{
     @Override
     public void setPassword(String password) {
         this.password = password;
-    }
-
-    @Override
-    public void setSecurityLevel(int securityLevel) {
-        this.securityLevel = securityLevel;
     }
 }
